Switch between paper/spigot events
