+++ .gitignore
@@ -1,11 +1,6 @@
-#Maven
-**/target
-dependency-reduced-pom.xml
-
 #IntelliJ
 .idea
 *.iml
-crazycrates/META-INF/
 
 #Eclipse
 **.settings
@@ -15,13 +10,8 @@ crazycrates/META-INF/
 \.DS_Store
 
 # Gradle
-api/build
+gradlew
 gradlew.bat
-plugin/build
-
-plugin/src/main/java/com/badbones69/crazycrates/func/v2/data
-build
 
 .gradle
-gradle
-gradlew
+build
